move vector geometry to core
